requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email-validator==2.3.0
fastapi==0.111.0
fastapi-cli==0.0.10
fastapi-mcp @ git+https://github.com/tadata-org/fastapi_mcp.git@e5cad13cabfc725bbcb047e526816d887d96da62
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mcp==1.12.4
mdurl==0.1.2
orjson==3.11.3
packaging==25.0
pluggy==1.6.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.2
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rpds-py==0.27.1
setuptools==80.9.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.37.2
stix2==3.0.1
stix2-patterns==2.0.0
taxii2-client==2.3.0
tomli==2.2.1
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.30.1
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
